sentiment_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2==2.9.9
transformers==4.36.2
torch==2.1.2
scipy==1.11.4
python-dotenv==1.0.0